+++ EpicJewels/GemEffects/BurningFrenzy.cs
@@ -23,11 +23,11 @@ namespace EpicJewels.GemEffects
                 if (hit.GetAttacker() is Player)
                 {
                     bool thisVikingisOnFIRE = Player.m_localPlayer.GetSEMan().HaveStatusEffect(burningstatus);
-                    if (Common.Config.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"Frenzy checking for burning viking {thisVikingisOnFIRE}"); }
+                    EpicJewels.EJLog.LogDebug($"Frenzy checking for burning viking {thisVikingisOnFIRE}");
                     if (thisVikingisOnFIRE)
                     {
                         float onfire_damage_bonus = (100 + Player.m_localPlayer.GetEffectPower<Config>("BurningFrenzy").Power) / 100;
-                        if (Common.Config.EnableDebugMode.Value) { Jotunn.Logger.LogInfo($"VikingOnFire Damage multiplier {onfire_damage_bonus}"); }
+                        EpicJewels.EJLog.LogDebug($"VikingOnFire Damage multiplier {onfire_damage_bonus}");
                         hit.m_damage.m_blunt *= onfire_damage_bonus;
                         hit.m_damage.m_pierce *= onfire_damage_bonus;
                         hit.m_damage.m_pierce *= onfire_damage_bonus;