+++ EpicJewels/GemEffects/CoinGreed.cs
@@ -9,6 +9,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
+using static CharacterDrop;
 
 namespace EpicJewels.GemEffects
 {
@@ -22,24 +23,31 @@ namespace EpicJewels.GemEffects
             //[InverseMultiplicativePercentagePower] public float Chance;
         }
 
-        [HarmonyPatch(typeof(CharacterDrop), nameof(CharacterDrop.GenerateDropList))]
-        public static class AddGreedToCreatureDrop
+        [HarmonyPatch(typeof(Character), nameof(Character.OnDeath))]
+        public static class AddGreedOnDeathFromPlayer
         {
             [UsedImplicitly]
-            private static void Postfix(CharacterDrop __instance, ref List<KeyValuePair<GameObject, int>> __result)
+            private static void Postfix(Character __instance)
             {
+                if (__instance.m_lastHit.GetAttacker() != Player.m_localPlayer)
+                {
+                    return;
+                }
                 float coin_greed = Player.m_localPlayer.GetEffectPower<Config>("CoinGreed").Power;
-                Jotunn.Logger.LogInfo($"Coingreed enabled? {coin_greed > 0}");
+                EpicJewels.EJLog.LogDebug($"Coingreed enabled? {coin_greed > 0}");
                 if (coin_greed > 0)
                 {
                     float greedChance = 0.15f;
                     float chance_roll = UnityEngine.Random.value;
-                    Jotunn.Logger.LogInfo($"Coingreed: {chance_roll} < {greedChance} = {chance_roll < greedChance}");
+                    EpicJewels.EJLog.LogDebug($"Coingreed: {chance_roll} < {greedChance} = {chance_roll < greedChance}");
                     if (chance_roll < greedChance)
                     {
                         float greedAmount = UnityEngine.Random.Range(1, Math.Max(1, coin_greed));
-                        Jotunn.Logger.LogInfo($"coingreed drop {greedAmount}");
-                        __result.Add(new KeyValuePair<GameObject, int>(ObjectDB.instance.GetItemPrefab("Coins"), (int)greedAmount));
+                        EpicJewels.EJLog.LogDebug($"coingreed drop {(int)greedAmount}");
+                        for (int i = 0; i <= (int)greedAmount; i++)
+                        {
+                            UnityEngine.Object.Instantiate(ObjectDB.instance.GetItemPrefab("Coins"), __instance.gameObject.transform.position, __instance.gameObject.transform.rotation);
+                        }
                     }
                 }