+++ EpicJewels/Package/CHANGELOG.md
@@ -1,3 +1,19 @@
+ **0.2.1**
+---
+```
+- Fixes stamina consumption for WeaponReducedStamina
+```
+
+ **0.2.0**
+---
+```
+- Added CoverOfDarkness to Opal
+- Removes Jotunn dependency
+- Changes coingreed to apply only to character deaths killed by the player
+- Removed excess logging
+- Removed debug logging option
+```
+
  **0.1.0**
 ---
 ```